refactor(service-api-endpoint): extract auth/middleware/payload/server/websocket modules (#5215)
